* add confirm dialog when remove entry shortcut.
git-svn-id: file:///tmp/ranzhi/trunk@2290 cb094fd3-5a31-0410-bd9d-ef5a9cbf4261

+++ www/js/jquery/ips.js
@@ -708,7 +708,7 @@
             if(!$.isFunction(settings.onDeleteEntry)) return;
             var id = $(this).closest('.dropdown-menu').data('id');
             var et = getEntry(id);
-            if(et)
+            if(et && confirm(settings.confirmRemoveEntry.format(et.name)))
             {
                 settings.onDeleteEntry(et, function(result)
                 {
